flask_backend/requirements.txt
flask
flask-cors
tensorflow
numpy
Pillow
